requirements.txt
# Core
python-dotenv>=1.0.0

# Document loaders
pypdf>=4.0.0
python-docx>=1.1.0
python-pptx>=0.6.23
pandas>=2.0.0
openpyxl>=3.1.0

# Chunking & retrieval
langchain-text-splitters>=0.0.1
langchain-core>=0.1.0
rank-bm25>=0.2.2
sentence-transformers>=2.5.0
faiss-cpu>=1.7.4
numpy>=1.24.0

# LLM providers
langchain-ollama>=0.1.0
openai>=1.12.0
# Optional (uncomment if using OpenAI as provider):
# langchain-openai>=0.1.0

# UI
streamlit>=1.31.0
